Rewrite clifton00multijava entry
BibTeX complained (because the entry wasn't @article):

> Warning--can't use both volume and number fields in clifton00multijava

but in fact the whole record was bad (as a rule, I avoid citing SIGPLAN
Notices -- either there's a conference entry to cite, or the paper is
not from a conference, was not peer-reviewed and is usually not worth
reading).

I tried following Tillmann's format, though it seems overkill.